fix(help): address lint and CodeRabbit review
- Extract helpSections() into internal/ui/help_sections.go so help.go
  drops back under the 800-line repo cap (revive file-length-limit).
- Preallocate widths slice in TestBuildHelpSpecs_KeysAlignWithinSection
  and reuse capacity across sections (golangci prealloc).
- docs/keybindings.md: S row no longer claims save_resource — that
  config key belongs to W. S maps to scale only.
