+++ internal/ui/help_sections.go
@@ -1,0 +1,474 @@
+package ui
+
+// helpSections returns all help sections with their keybindings.
+//
+// Lives in its own file so the static catalog can grow without
+// pushing help.go past the repo's 800-line file cap.
+func helpSections() []helpSection {
+	kb := ActiveKeybindings
+	return []helpSection{
+		{
+			title: "Navigation",
+			bindings: []helpEntry{
+				{kb.Left + " / Left", "Go to parent"},
+				{kb.Right + " / Right", "Go to child"},
+				{kb.Down + " / Down", "Move down"},
+				{kb.Up + " / Up", "Move up"},
+				{kb.JumpTop + kb.JumpTop + " / Home", "Jump to top"},
+				{kb.JumpBottom + " / End", "Jump to bottom"},
+				{helpKeyDisplay(kb.PageDown) + " / " + helpKeyDisplay(kb.PageUp), "Half-page scroll down / up"},
+				{helpKeyDisplay(kb.PageForward) + " / " + helpKeyDisplay(kb.PageBack) + " / PgDn / PgUp", "Full-page scroll down / up"},
+				{kb.Enter, "Open YAML view / navigate into"},
+				{kb.LevelCluster + "/" + kb.LevelTypes + "/" + kb.LevelResources, "Jump to clusters/types/resources level"},
+				{kb.JumpOwner, "Jump to owner/controller of selected resource"},
+				{kb.PreviewDown + "/" + kb.PreviewUp, "Scroll preview pane down/up"},
+				{kb.ExpandCollapse, "Toggle expand/collapse all resource groups / toggle event grouping (Events)"},
+				{kb.PinGroup, "Pin/unpin CRD group (at resource types level)"},
+				{kb.ToggleRare, "Toggle rarely used resource types (CSI, webhooks, advanced) in the sidebar"},
+			},
+		},
+		{
+			title: "Search & Filter",
+			bindings: []helpEntry{
+				{kb.Filter, "Filter items (~prefix: fuzzy, regex auto-detected, \\prefix: literal)"},
+				{kb.Search, "Search and jump to match (~fuzzy, regex auto, \\literal)"},
+				{"", "Up/Down inside filter or search recalls previous queries (shared, persistent across sessions)."},
+				{"", "Paste from clipboard: Cmd+V (macOS) / Ctrl+Shift+V (Linux). Multiline asks to confirm."},
+				{kb.NextMatch, "Next search match"},
+				{kb.PrevMatch, "Previous search match"},
+				{kb.FilterPresets, "Quick filter presets"},
+			},
+		},
+		{
+			title: "Views & Tools",
+			bindings: []helpEntry{
+				{kb.Help + " / F1", "Toggle help screen"},
+				{kb.TogglePreview, "Toggle between details and YAML preview"},
+				{"", "Details pane shows labels, finalizers, annotations, and metadata"},
+				{"", "Details view shows deletion timestamp for resources being deleted"},
+				{kb.Fullscreen, "Toggle fullscreen (middle column or dashboard)"},
+				{kb.ResourceMap, "Toggle resource relationship map"},
+				{kb.ColumnToggle, "Column visibility toggle (show/hide and reorder columns)"},
+				{helpKeyDisplay(kb.SecretToggle), "Toggle secret value visibility (details pane only)"},
+				{kb.APIExplorer, "API Explorer (resource structure)"},
+				{kb.RBACBrowser, "RBAC permissions browser (can-i)"},
+				{helpKeyDisplay(kb.FinalizerSearch), "Finalizer search and remove (scan, select, remove finalizers)"},
+				{kb.ErrorLog, "Error log"},
+				{kb.Monitoring, "Monitoring overview (active Prometheus alerts)"},
+				{kb.QuotaDashboard, "Namespace resource quota dashboard"},
+				{kb.TasksOverlay, "Background tasks overlay (Tab toggles running/completed history)"},
+			},
+		},
+		{
+			title: "Sorting",
+			bindings: []helpEntry{
+				{kb.SortNext, "Sort by next column"},
+				{kb.SortPrev, "Sort by previous column"},
+				{kb.SortFlip, "Toggle sort direction (ascending/descending)"},
+				{kb.SortReset, "Reset sort to default (Name ascending)"},
+			},
+		},
+		{
+			title: "Multi-Selection",
+			bindings: []helpEntry{
+				{"Space", "Toggle selection on current item"},
+				{"Ctrl+Space", "Select range from anchor to cursor"},
+				{helpKeyDisplay(kb.SelectAll), "Select/deselect all visible items"},
+				{"Esc", "Clear selection"},
+				{kb.ActionMenu, "Bulk action menu (delete, force delete, scale, restart, diff, ArgoCD sync/refresh)"},
+				{kb.Diff, "Diff: compare YAML of two selected resources"},
+			},
+		},
+		{
+			title: "Actions",
+			bindings: []helpEntry{
+				{kb.NamespaceSelector, "Select namespace"},
+				{kb.AllNamespaces, "Toggle all-namespaces mode"},
+				{kb.ActionMenu, "Action menu: l=tail logs (last 10 lines + follow), L=full logs, exec, debug, debug pod, describe, edit, delete, scale, port-forward, events, startup analysis, RBAC permissions"},
+				{kb.Logs, "View full logs for selected resource"},
+				{kb.Describe, "Describe selected resource"},
+				{kb.Edit, "Edit selected resource in $EDITOR"},
+				{kb.SecretEditor, "Secret/ConfigMap editor (inline key-value editing)"},
+				{kb.LabelEditor, "Edit labels/annotations for selected resource"},
+				{kb.Refresh, "Refresh current view"},
+				{kb.Delete, "Delete (force delete Pod/Job if already deleting, force finalize others)"},
+				{kb.ForceDelete, "Force delete (Pod/Job only)"},
+				{kb.SaveResource, "Save resource to file / toggle warnings-only (Events)"},
+				{kb.CopyName, "Copy resource name to clipboard"},
+				{helpKeyDisplay(kb.CopyYAML), "Copy resource YAML to clipboard"},
+				{helpKeyDisplay(kb.PasteApply), "Apply resource from clipboard (kubectl apply)"},
+				{"", "Port forwarding: use action menu (" + kb.ActionMenu + ") on Pod/Service/Deployment/StatefulSet/DaemonSet"},
+				{"", "Auto-navigates to Port Forwards list after creation; shows resolved local port"},
+			},
+		},
+		{
+			title: "Modes & Settings",
+			bindings: []helpEntry{
+				{kb.WatchMode, "Toggle watch mode (auto-refresh)"},
+				{helpKeyDisplay(kb.ReadOnlyToggle), "Toggle read-only mode (cluster picker: highlighted row's [RO] marker; inside a context: current tab)"},
+				{helpKeyDisplay(kb.ClusterColorPicker), "Cluster color picker (cluster picker only): tint title bar + add row swatch; persisted across restarts"},
+				{kb.ThemeSelector, "Switch color scheme (" + kb.NewTab + ": toggle transparent bg)"},
+				{helpKeyDisplay(kb.TerminalToggle), "Cycle terminal mode (pty / exec / mux — mux skipped if no tmux/zellij detected)"},
+			},
+		},
+		{
+			title: "Command Bar (:)",
+			bindings: []helpEntry{
+				{kb.CommandBar, "Open command bar"},
+				{"", "Resource jump: :pod, :dep, :svc (navigate to resource type)"},
+				{"", "  With namespace: :pod kube-system (jump + filter namespace)"},
+				{"", "Built-in: :ns (navigate to NS or filter), :ctx <name>, :set <opt>, :sort <col>, :export <fmt>"},
+				{"", "Kubectl: :k get pod, :kubectl describe pod (requires k/kubectl prefix)"},
+				{"", "Shell: :! <command> (run arbitrary shell command)"},
+				{"", ""},
+				{"Tab", "Cycle suggestions forward (auto-fill when 1 match)"},
+				{"Shift+Tab", "Cycle suggestions backward"},
+				{"Ctrl+N / Down", "Cycle suggestions forward"},
+				{"Ctrl+P / Up", "Cycle suggestions backward"},
+				{"Ctrl+D", "Scroll suggestions down (half page)"},
+				{"Ctrl+U", "Scroll suggestions up (half page) / delete line when closed"},
+				{"Ctrl+F / Ctrl+B", "Scroll suggestions (full page)"},
+				{"Ctrl+Space", "Open/refresh suggestions"},
+				{"Space / Right", "Accept ghost text preview"},
+				{"Enter", "Accept suggestion (if visible) or execute command"},
+				{"Esc", "Close suggestions, or close command bar"},
+				{"Up / Down", "Browse command history (when no suggestions)"},
+				{"Ctrl+W", "Delete word backwards"},
+				{"Ctrl+A / Ctrl+E", "Home / End"},
+			},
+		},
+		{
+			title: "Bookmarks",
+			bindings: []helpEntry{
+				{kb.SetMark + "<a-z/0-9>", "Set context-aware mark (switches cluster on jump)"},
+				{kb.SetMark + "<A-Z>", "Set context-free mark (stays in current cluster on jump)"},
+				{kb.OpenMarks, "Open bookmarks list"},
+				{"a-z/A-Z/0-9", "Jump to named mark (in overlay)"},
+				{"j/k", "Navigate bookmarks (in overlay)"},
+				{"/", "Filter bookmarks by name (in overlay)"},
+				{"Enter", "Jump to selected bookmark (in overlay)"},
+				{"Ctrl+X", "Delete selected bookmark with confirmation (in overlay)"},
+				{"Alt+X", "Delete all bookmarks with confirmation (in overlay)"},
+			},
+		},
+		{
+			title: "Error Log (" + kb.ErrorLog + ")", context: "Error Log",
+			bindings: []helpEntry{
+				{"j/k", "Scroll up/down"},
+				{"gg/G / Home/End", "Top/bottom"},
+				{"Ctrl+D/U", "Page down/up (half page)"},
+				{"Ctrl+F/B / PgDn/PgUp", "Page down/up (full page)"},
+				{"V", "Enter line visual selection mode"},
+				{"v", "Enter character visual selection mode"},
+				{"y", "Copy selected lines (visual) or all entries (normal)"},
+				{"f", "Toggle fullscreen / overlay mode"},
+				{"d", "Toggle debug log visibility"},
+				{"Esc", "Cancel visual selection, or close overlay"},
+				{"q", "Close overlay"},
+			},
+		},
+		{
+			title: "YAML View", context: "YAML View",
+			bindings: []helpEntry{
+				{"j/k", "Scroll up/down"},
+				{"123j/123k", "Move cursor down/up N visible lines (count-prefixed; folds skipped)"},
+				{"h/l", "Move cursor column left/right"},
+				{"123h/123l", "Move cursor column left/right by N runes"},
+				{"0/$", "Move cursor to line start/end"},
+				{"^", "Move cursor to first non-whitespace"},
+				{"w/b", "Move cursor to next/previous word start"},
+				{"W/B", "Move cursor to next/previous WORD start"},
+				{"e", "Move cursor to end of word"},
+				{"E", "Move cursor to end of WORD"},
+				{"123w/123b/123e (and capitals)", "Apply word/WORD motion N times"},
+				{"gg/G / Home/End", "Top/bottom"},
+				{"123G", "Jump to line number"},
+				{"Ctrl+D/U", "Page down/up (half page)"},
+				{"Ctrl+F/B / PgDn/PgUp", "Page down/up (full page)"},
+				{"123 Ctrl+D/U", "Scroll N lines (sticky vim 'scroll', shared C-d/C-u)"},
+				{"123 Ctrl+F/B", "Page motion scaled by N"},
+				{"/", "Search in YAML"},
+				{"n", "Next search match"},
+				{"N", "Previous search match"},
+				{"123n/123N", "Jump to Nth next/previous match"},
+				{"v", "Character visual selection (from cursor column)"},
+				{"V", "Visual line selection"},
+				{"Ctrl+V", "Block (column) visual selection (from cursor column)"},
+				{"h/l", "Move selection column (in visual mode)"},
+				{"y", "Copy line (or selection in visual mode)"},
+				{"123y", "Copy number of lines from cursor (count-prefixed yank, folds skipped)"},
+				{"z", "Toggle fold on section under cursor"},
+				{"Z", "Toggle all folds (collapse/expand)"},
+				{"Ctrl+W / >", "Toggle line wrapping"},
+				{"Ctrl+E", "Edit resource in editor"},
+				{"q/Esc", "Back to explorer"},
+			},
+		},
+		{
+			title: "Describe View", context: "Describe View",
+			bindings: []helpEntry{
+				{"j/k", "Move cursor up/down"},
+				{"123j/123k", "Move cursor down/up N lines (count-prefixed)"},
+				{"h/l", "Move cursor column left/right"},
+				{"123h/123l", "Move cursor column left/right by N runes"},
+				{"0/$", "Move cursor to line start/end"},
+				{"^", "Move cursor to first non-whitespace"},
+				{"w/b", "Move cursor to next/previous word start"},
+				{"W/B", "Move cursor to next/previous WORD start"},
+				{"e", "Move cursor to end of word"},
+				{"E", "Move cursor to end of WORD"},
+				{"123w/123b/123e (and capitals)", "Apply word/WORD motion N times"},
+				{"gg/G / Home/End", "Top/bottom"},
+				{"123G", "Jump to line number"},
+				{"Ctrl+D/U", "Page down/up (half page)"},
+				{"Ctrl+F/B / PgDn/PgUp", "Page down/up (full page)"},
+				{"123 Ctrl+D/U", "Scroll N lines (sticky vim 'scroll', shared C-d/C-u)"},
+				{"123 Ctrl+F/B", "Page motion scaled by N"},
+				{"/", "Search in content"},
+				{"n/N", "Next/previous match"},
+				{"123n/123N", "Jump to Nth next/previous match"},
+				{"v", "Character visual selection"},
+				{"V", "Visual line selection"},
+				{"Ctrl+V", "Block (column) visual selection"},
+				{"y", "Copy line (or selection in visual mode)"},
+				{"123y", "Copy number of lines from cursor (count-prefixed yank)"},
+				{"Ctrl+W / >", "Toggle line wrapping"},
+				{"q/Esc", "Back to explorer"},
+			},
+		},
+		{
+			title: "Diff View", context: "Diff View",
+			bindings: []helpEntry{
+				{"j/k", "Move cursor up/down"},
+				{"123j/123k", "Move cursor down/up N lines (count-prefixed)"},
+				{"h/l", "Move cursor column left/right"},
+				{"123h/123l", "Move cursor column left/right by N runes"},
+				{"0/$", "Move cursor to line start/end"},
+				{"^", "Move cursor to first non-whitespace"},
+				{"w/b", "Move cursor to next/previous word start"},
+				{"W/B", "Move cursor to next/previous WORD start"},
+				{"e", "Move cursor to end of word"},
+				{"E", "Move cursor to end of WORD"},
+				{"123w/123b/123e (and capitals)", "Apply word/WORD motion N times"},
+				{"Tab", "Switch cursor side (side-by-side mode)"},
+				{"gg/G / Home/End", "Top/bottom"},
+				{"123G", "Jump to line number"},
+				{"Ctrl+D/U", "Page down/up (half page)"},
+				{"Ctrl+F/B / PgDn/PgUp", "Page down/up (full page)"},
+				{"123 Ctrl+D/U", "Scroll N lines (sticky vim 'scroll', shared C-d/C-u)"},
+				{"123 Ctrl+F/B", "Page motion scaled by N"},
+				{"/", "Search in diff"},
+				{"n/N", "Next/previous match"},
+				{"123n/123N", "Jump to Nth next/previous match"},
+				{"v", "Character visual selection"},
+				{"V", "Visual line selection"},
+				{"Ctrl+V", "Block (column) visual selection"},
+				{"h/l", "Move selection column (in visual mode)"},
+				{"y", "Copy line (or selection in visual mode)"},
+				{"123y", "Copy number of lines from cursor (count-prefixed yank, empty rows skipped)"},
+				{"z", "Toggle fold unchanged section at cursor"},
+				{"Z", "Toggle all folds"},
+				{"#", "Toggle line numbers"},
+				{"Ctrl+W / >", "Toggle line wrapping"},
+				{"u", "Toggle unified/side-by-side view"},
+				{"q/Esc", "Back to explorer"},
+			},
+		},
+		{
+			title: "API Explorer", context: "API Explorer",
+			bindings: []helpEntry{
+				{"j/k", "Navigate fields"},
+				{"l/Enter", "Drill into field (Object/array types)"},
+				{"h/Backspace", "Go back one level"},
+				{"/", "Search fields"},
+				{"n", "Next match (auto-drills into children)"},
+				{"N", "Previous match (searches parent)"},
+				{"r", "Recursive field browser (browse all nested fields with filter)"},
+				{"gg/G / Home/End", "Top/bottom"},
+				{"Ctrl+D/U", "Page down/up (half page)"},
+				{"Ctrl+F/B / PgDn/PgUp", "Page down/up (full page)"},
+				{"q", "Close API explorer"},
+				{"Esc", "Go back one level / close at root"},
+			},
+		},
+		{
+			title: "Can-I Browser", context: "Can-I Browser",
+			bindings: []helpEntry{
+				{"j/k", "Navigate groups"},
+				{"J/K", "Scroll resource list down/up"},
+				{"/", "Search/filter groups by name"},
+				{"a", "Toggle all/allowed-only permissions"},
+				{"s", "Switch subject (User/Group/SA)"},
+				{"", "Title shows namespace scope (ns:...) for permission context"},
+				{"gg/G / Home/End", "Top/bottom"},
+				{"Ctrl+D/U", "Page down/up (half page)"},
+				{"Ctrl+F/B / PgDn/PgUp", "Page down/up (full page)"},
+				{"q/Esc", "Clear search / close"},
+			},
+		},
+		{
+			title: "Can-I Subject Selector", context: "Can-I Browser",
+			bindings: []helpEntry{
+				{"j/k", "Navigate subjects"},
+				{"/", "Filter subjects by name"},
+				{"gg/G / Home/End", "Top/bottom"},
+				{"Ctrl+D/U", "Page down/up (half page)"},
+				{"Ctrl+F/B / PgDn/PgUp", "Page down/up (full page)"},
+				{"Enter", "Select subject"},
+				{"Esc", "Clear filter / close"},
+			},
+		},
+		{
+			title: "Network Policy Visualizer", context: "Network Policy",
+			bindings: []helpEntry{
+				{"j/k", "Scroll up/down"},
+				{"gg/G / Home/End", "Top/bottom"},
+				{"Ctrl+D/U", "Page down/up (half page)"},
+				{"Ctrl+F/B / PgDn/PgUp", "Page down/up (full page)"},
+				{"q/Esc", "Close visualizer"},
+			},
+		},
+		{
+			title: "Log Viewer", context: "Log Viewer",
+			bindings: []helpEntry{
+				{"j/k", "Move cursor up/down"},
+				{"123j/123k", "Move cursor down/up N lines (count-prefixed)"},
+				{"h/l/left/right", "Move cursor column left/right"},
+				{"123h/123l", "Move cursor column left/right by N runes"},
+				{"0/$", "Move cursor to line start/end"},
+				{"^", "Move cursor to first non-whitespace"},
+				{"w/b", "Move cursor to next/previous word start"},
+				{"W/B", "Move cursor to next/previous WORD start"},
+				{"e", "Move cursor to end of word"},
+				{"E", "Move cursor to end of WORD"},
+				{"123w/123b/123e (and capitals)", "Apply word/WORD motion N times"},
+				{"gg/G / Home/End", "Top/bottom"},
+				{"Ctrl+D/U", "Half page down/up"},
+				{"Ctrl+F/B / PgDn/PgUp", "Full page down/up"},
+				{"123 Ctrl+D/U", "Scroll N lines (sticky vim 'scroll', shared C-d/C-u)"},
+				{"123 Ctrl+F/B", "Page motion scaled by N"},
+				{"f", "Toggle follow mode (auto-scroll)"},
+				{"Tab/z/>", "Toggle line wrapping"},
+				{"#", "Toggle line numbers"},
+				{"s", "Toggle timestamps"},
+				{"p", "Toggle pod/container prefixes"},
+				{"P", "Toggle structured preview side panel (JSON / logfmt / klog / zap / nginx / envoy / java / postgres / text)"},
+				{"J/K", "Scroll preview side panel down/up (when visible)"},
+				{"c", "Toggle previous container logs"},
+				{"/", "Search in logs"},
+				{"", "Up/Down inside log search recalls previous queries (persistent across sessions)."},
+				{"n", "Next search match"},
+				{"N", "Previous search match"},
+				{"123n/123N", "Jump to Nth next/previous match"},
+				{"123G", "Jump to line number"},
+				{"S", "Save loaded logs to file (path copied to clipboard)"},
+				{"Ctrl+S", "Save all logs to file (path copied to clipboard)"},
+				{"v", "Character visual selection (from cursor column)"},
+				{"V", "Visual line selection"},
+				{"Ctrl+V", "Block (column) visual selection (from cursor column)"},
+				{"h/l", "Move selection column (in visual mode)"},
+				{"y", "Copy line (or selection in visual mode)"},
+				{"123y", "Copy number of lines from cursor (count-prefixed yank)"},
+				{"\\", "Switch pod / filter containers"},
+				{"", "Full logs load last 1000 lines initially (log_tail_lines); tail logs load last 10 (log_tail_lines_short). Scroll up for older history."},
+				{"q/Esc", "Close log viewer"},
+			},
+		},
+		{
+			title: "Event Timeline", context: "Event Timeline",
+			bindings: []helpEntry{
+				{"j/k", "Move cursor down/up"},
+				{"123j/123k", "Move cursor down/up N lines (count-prefixed)"},
+				{"h/l/left/right", "Move cursor column left/right"},
+				{"123h/123l", "Move cursor column left/right by N runes"},
+				{"0/$", "Move cursor to line start/end"},
+				{"^", "Move cursor to first non-whitespace"},
+				{"w/b", "Move cursor to next/previous word start"},
+				{"W/B", "Move cursor to next/previous WORD start"},
+				{"e", "Move cursor to end of word"},
+				{"E", "Move cursor to end of WORD"},
+				{"123w/123b/123e (and capitals)", "Apply word/WORD motion N times"},
+				{"gg/G / Home/End", "Top/bottom"},
+				{"123G", "Jump to line number"},
+				{"Ctrl+D/U", "Half page down/up"},
+				{"Ctrl+F/B / PgDn/PgUp", "Full page down/up"},
+				{"123 Ctrl+D/U", "Scroll N lines (sticky vim 'scroll', shared C-d/C-u)"},
+				{"123 Ctrl+F/B", "Page motion scaled by N"},
+				{"f", "Toggle fullscreen event viewer"},
+				{"Tab/z/>", "Toggle line wrapping"},
+				{"/", "Search in events"},
+				{"n", "Next search match"},
+				{"N", "Previous search match"},
+				{"123n/123N", "Jump to Nth next/previous match"},
+				{"v", "Character visual selection (from cursor column)"},
+				{"V", "Visual line selection"},
+				{"Ctrl+V", "Block (column) visual selection (from cursor column)"},
+				{"y", "Copy line (or selection in visual mode)"},
+				{"123y", "Copy number of lines from cursor (count-prefixed yank)"},
+				{"", "Events are correlated to the selected resource and grouped when Type/Reason/Message/Object match. The 1-9 line buffer is consumed after each motion so digits do not leak between commands."},
+				{"q/Esc", "Close overlay (or exit fullscreen back to overlay)"},
+			},
+		},
+		{
+			title: "Exec Mode (embedded terminal)", context: "Exec Mode",
+			bindings: []helpEntry{
+				{"Ctrl+]", "Prefix key (like tmux Ctrl+b)"},
+				{"Ctrl+] -> Ctrl+]", "Exit terminal and return to explorer"},
+				{"Ctrl+] -> " + kb.NextTab, "Next tab (PTY keeps running in background)"},
+				{"Ctrl+] -> " + kb.PrevTab, "Previous tab (PTY keeps running in background)"},
+				{"Ctrl+] -> " + kb.NewTab, "New tab (clone current context)"},
+				{"Ctrl+] -> Ctrl+U / Ctrl+D", "Scroll back / forward by half a viewport"},
+				{"Ctrl+] -> Ctrl+B / Ctrl+F", "Scroll back / forward by a full viewport"},
+				{"Ctrl+] -> g / G", "Jump to oldest captured line / back to live"},
+				{"Mouse Scroll", "Scroll the PTY scrollback (1 line per tick)"},
+				{"Shift+Drag", "Select text (host terminal); Shift+Option+Drag for block select on macOS"},
+				{"", "Typing any key forwards to the PTY and snaps back to the live shell"},
+				{"", "Scrollback is captured from the PTY byte stream (newline-split, ANSI-stripped, ~5000 lines). Curses programs (vim, less, htop) that paint absolute screen positions will produce noisy scrollback while running."},
+				{"", "Cycle terminal mode with Ctrl+T: pty (embedded) -> exec (host terminal) -> mux (new tmux/zellij window or pane). Mux is skipped when no multiplexer is detected. Set terminal: pty|exec|mux in config to pick the default."},
+			},
+		},
+		{
+			title: "Tabs",
+			bindings: []helpEntry{
+				{kb.NewTab, "New tab (clone current)"},
+				{kb.PrevTab, "Previous tab"},
+				{kb.NextTab, "Next tab"},
+			},
+		},
+		{
+			title: "Mouse",
+			bindings: []helpEntry{
+				{"Click", "Select item / navigate"},
+				{"Scroll", "Navigate up/down"},
+				{"Shift+Drag", "Select text (terminal native)"},
+			},
+		},
+		{
+			title: "Help View",
+			bindings: []helpEntry{
+				{"/", "Search — highlights matches inline without filtering"},
+				{"Ctrl+N / Ctrl+P", "Next / previous match while typing the search"},
+				{"Enter", "Apply search (keep highlights and arm n/N)"},
+				{"n / N", "Jump to next / previous search match (after Enter)"},
+				{"f", "Filter — narrows the visible list to matching lines"},
+				{"Esc", "Cascades: clear search → clear filter → close help"},
+			},
+		},
+		{
+			title: "General",
+			bindings: []helpEntry{
+				{"q", "Quit (with confirmation)"},
+				{"Esc", "Go back / quit"},
+				{"Ctrl+C", "Close tab (quit if last)"},
+			},
+		},
+		{
+			title: "Configuration",
+			bindings: []helpEntry{
+				{"", "Config: ~/.config/lfk/config.yaml (or $XDG_CONFIG_HOME/lfk/config.yaml)"},
+				{"", "State:  ~/.local/state/lfk/ (bookmarks, session, history, pinned groups)"},
+				{"", "Logs:   ~/.local/share/lfk/lfk.log"},
+			},
+		},
+	}
+}
